+++ src/com/intellij/idea/plugin/hybris/toolwindow/ccv2/views/CCv2UiElements.kt
@@ -67,3 +67,10 @@ fun Panel.ccv2ServiceModifiedByRow(service: CCv2ServiceDto) {
             .comment("Modified by")
     }
 }
+
+fun Panel.ccv2ServiceModifiedTimeRow(service: CCv2ServiceDto) {
+    row {
+        label(service.modifiedTimeFormatted)
+            .comment("Modified time")
+    }
+}