+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2024.1.3]
 
+### `CCv2` enhancements
+- Show modified time for Service view [#1127](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1127)
+
 ### Other
 - Migrated to `IntelliJ Platform Gradle Plugin 2.beta2` [#1124](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1124)
 - Show once `CHANGELOG.md` in a new `What's New` tab once per version [#1125](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1125)